# backend/account-banking-api/src/main/resources/openapi/account-banking-api.yaml
openapi: 3.0.3
info:
  title: Account Banking API
  version: 0.1.0
  description: Account context API for Banking Simulation Platform.
paths:
  /accounts:
    post:
      summary: Create a bank account
      responses:
        '201':
          description: Account created
    get:
      summary: List accounts by owner
      parameters:
        - name: owner_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 25
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Accounts found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsResponse'
  /accounts/{accountId}:
    get:
      summary: Get account by id
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Account not found
components:
  schemas:
    AccountsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AccountResponse'
        limit:
          type: integer
        offset:
          type: integer
        nextOffset:
          type: integer
    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        iban:
          type: string
        status:
          type: string
        balance:
          $ref: '#/components/schemas/MoneyResponse'
    MoneyResponse:
      type: object
      properties:
        amount:
          type: string
        currency:
          type: string